requirements.txt
# platform: linux-64
numpy==1.26.4
pandas==2.1.1
scikit-learn==1.3.2
biopython==1.83
viennarna==2.6.4
